class exercise solution

+++ src/database.test.ts
@@ -6,20 +6,28 @@ describe('Database', () => {
 
     beforeEach(() => {
         // clean out  __dirname/documents
+
+
+        database = new Database();
+    });
+
+    afterAll(() => {
         const fs = require('fs');
         const path = require('path');
         const directory = path.join(__dirname, 'documents');
         const files = fs.readdirSync(directory);
+
         for (const file of files) {
-            // delete the file
-            fs.unlinkSync(path.join(directory, file));
+            // check if it is a test file
+            if (file.startsWith('test_xxx')) {
+                fs.unlinkSync(path.join(directory, file));
+            }
         }
-        database = new Database();
     });
 
     describe('getDocument', () => {
         it('should return a document with the given name', () => {
-            const documentName = 'test';
+            const documentName = 'test_xxx';
             const document = database.createDocument(documentName);
             // add a task to the document
             document.addTask('test task');
@@ -31,15 +39,12 @@ describe('Database', () => {
             expect(result).toBe(document);
         });
 
-        it('should throw an error if no document with the given name exists', () => {
-            expect(() => database.getDocument('nonexistent')).toThrowError('Document nonexistent does not exist');
-        });
     });
 
     describe('getDocuments', () => {
         it('should return an array of all documents in the database', () => {
-            const document1 = database.createDocument('test1');
-            const document2 = database.createDocument('test2');
+            const document1 = database.createDocument('test_xxx1');
+            const document2 = database.createDocument('test_xxx2');
             const result = database.getDocuments();
             expect(result).toContain(document1);
             expect(result).toContain(document2);
@@ -48,7 +53,7 @@ describe('Database', () => {
 
     describe('createDocument', () => {
         it('should create a new document with the given name', () => {
-            const documentName = 'test';
+            const documentName = 'test_xxx2';
             const fs = require('fs');
             const path = require('path');
             const file = path.join(__dirname, 'documents', documentName + '.json');
@@ -68,17 +73,11 @@ describe('Database', () => {
             expect(fileExists).toBe(true);
         });
 
-        // it('should return an existing document if one with the same name already exists', () => {
-        //     const documentName = 'test';
-        //     const document1 = database.createDocument(documentName);
-        //     const document2 = database.createDocument(documentName);
-        //     expect(document2).toBe(document1);
-        // });
     });
 
     describe('getTasks', () => {
         it('should return a map of tasks for the given document', () => {
-            const documentName = 'test';
+            const documentName = 'test_xxx45';
             const document = database.createDocument(documentName);
             document.addTask('test task');
             const result = database.getTasks(documentName);
@@ -89,17 +88,12 @@ describe('Database', () => {
 
         });
 
-        it('should create a new document if one with the given name does not exist', () => {
-            const documentName = 'nonexistent';
-            const result = database.getTasks(documentName);
-            expect(result.size).toBe(0);
-            expect(database.getDocument(documentName)).toBeDefined();
-        });
+
     });
 
     describe('reset', () => {
         it('should reset the tasks for the given document', () => {
-            const documentName = 'test';
+            const documentName = 'test_xxx3848';
             const document = database.createDocument(documentName);
             document.addTask('test task');
             database.reset(documentName);
@@ -107,31 +101,22 @@ describe('Database', () => {
             expect(result.size).toBe(0);
         });
 
-        it('should create a new document if one with the given name does not exist', () => {
-            const documentName = 'nonexistent';
-            database.reset(documentName);
-            expect(database.getDocument(documentName)).toBeDefined();
-        });
+
     });
 
     describe('makeData', () => {
         it('should populate the tasks for the given document', () => {
-            const documentName = 'test';
+            const documentName = 'test_xxx884383';
             database.makeData(documentName);
             const result = database.getTasks(documentName);
             expect(result.size).toBeGreaterThan(0);
         });
 
-        it('should create a new document if one with the given name does not exist', () => {
-            const documentName = 'nonexistent';
-            database.makeData(documentName);
-            expect(database.getDocument(documentName)).toBeDefined();
-        });
     });
 
     describe('addTask', () => {
         it('should add a task to the given document', () => {
-            const documentName = 'test';
+            const documentName = 'test_xxx88383';
             const result = database.addTask('test task', documentName);
             const tasks = database.getTasks(documentName);
             expect(tasks.size).toBe(1);
@@ -140,7 +125,18 @@ describe('Database', () => {
         });
 
         it('should create a new document if one with the given name does not exist', () => {
-            const documentName = 'nonexistent';
+            const documentName = 'test_xxx838479238838';
+            // check if the file exists
+            const fs = require('fs');
+            const path = require('path');
+            const file = path.join(__dirname, 'documents', documentName + '.json');
+
+            // delete the file if it exists
+            if (fs.existsSync(file)) {
+                fs.unlinkSync(file);
+            }
+
+
             const result = database.addTask('test task', documentName);
             const tasks = database.getTasks(documentName);
             expect(tasks.size).toBe(1);
@@ -152,7 +148,7 @@ describe('Database', () => {
 
     describe('deleteTask', () => {
         it('should delete the task with the given ID from the given document', () => {
-            const documentName = 'test';
+            const documentName = 'test_xxx3434';
             const taskId = database.addTask('test task', documentName);
             // assign the task to a user
             database.addUserToTask(taskId, 'user', documentName);
@@ -163,7 +159,7 @@ describe('Database', () => {
         });
 
         it('should not delete the task if the user is not assigned to it', () => {
-            const documentName = 'test';
+            const documentName = 'test_xxx12434234';
             const taskId = database.addTask('test task', documentName);
             // assign the task to a user
             database.addUserToTask(taskId, 'user', documentName);
@@ -174,17 +170,12 @@ describe('Database', () => {
             expect(tasks.size).toBe(1);
         });
 
-        it('should create a new document if one with the given name does not exist', () => {
-            const documentName = 'nonexistent';
-            const taskId = database.addTask('test task', documentName);
-            database.deleteTask(taskId, 'user', documentName);
-            expect(database.getDocument(documentName)).toBeDefined();
-        });
+
     });
 
     describe('addUserToTask', () => {
         it('should add the given user to the task with the given ID in the given document', () => {
-            const documentName = 'test';
+            const documentName = 'test_xxx2309478402789';
             const taskId = database.addTask('test task', documentName);
             const result = database.addUserToTask(taskId, 'user', documentName);
             const task = database.getTasks(documentName).get(taskId);
@@ -193,23 +184,16 @@ describe('Database', () => {
         });
 
         it('should return false if the task with the given ID does not exist in the given document', () => {
-            const documentName = 'test';
+            const documentName = 'test_xxxw988we8r7';
             const result = database.addUserToTask('nonexistent', 'user', documentName);
             expect(result).toBe(false);
         });
 
-        it('should create a new document if one with the given name does not exist', () => {
-            const documentName = 'nonexistent';
-            const taskId = database.addTask('test task', documentName);
-            const result = database.addUserToTask(taskId, 'user', documentName);
-            expect(database.getDocument(documentName)).toBeDefined();
-            expect(result).toBe(true);
-        });
     });
 
     describe('removeUserFromTask', () => {
         it('should remove the given user from the task with the given ID in the given document', () => {
-            const documentName = 'test';
+            const documentName = 'test_xxx2309478402789dds';
             const taskId = database.addTask('test task', documentName);
             database.addUserToTask(taskId, 'user', documentName);
             database.removeUserFromTask(taskId, 'user', documentName);
@@ -218,18 +202,11 @@ describe('Database', () => {
             expect(task?.owner).not.toContain('user');
         });
 
-        it('should create a new document if one with the given name does not exist', () => {
-            const documentName = 'nonexistent';
-            const taskId = database.addTask('test task', documentName);
-            database.addUserToTask(taskId, 'user', documentName);
-            database.removeUserFromTask(taskId, 'user', documentName);
-            expect(database.getDocument(documentName)).toBeDefined();
-        });
     });
 
     describe('addTimeToTask', () => {
         it('should add the given time to the task with the given ID and user in the given document', () => {
-            const documentName = 'test';
+            const documentName = 'test_xxx8838382';
             const taskId = database.addTask('test task', documentName);
             database.addUserToTask(taskId, 'user', documentName);
             const result = database.addTimeToTask(taskId, 'user', 10, documentName);
@@ -239,24 +216,15 @@ describe('Database', () => {
         });
 
         it('should return false if the task with the given ID does not exist in the given document', () => {
-            const documentName = 'test';
+            const documentName = 'test_xxxddkfr223';
             const result = database.addTimeToTask('nonexistent', 'user', 10, documentName);
             expect(result).toBe(false);
-        });
-
-        it('should create a new document if one with the given name does not exist', () => {
-            const documentName = 'nonexistent';
-            const taskId = database.addTask('test task', documentName);
-            database.addUserToTask(taskId, 'user', documentName);
-            const result = database.addTimeToTask(taskId, 'user', 10, documentName);
-            expect(database.getDocument(documentName)).toBeDefined();
-            expect(result).toBe(true);
         });
     });
 
     describe('markTaskComplete', () => {
         it('should mark the task with the given ID as complete for the given user in the given document', () => {
-            const documentName = 'test';
+            const documentName = 'test_xxx883838234';
             const taskId = database.addTask('test task', documentName);
             database.addUserToTask(taskId, 'user', documentName);
             const result = database.markTaskComplete(taskId, 'user', documentName);
@@ -266,18 +234,11 @@ describe('Database', () => {
         });
 
         it('should return false if the task with the given ID does not exist in the given document', () => {
-            const documentName = 'test';
+            const documentName = 'test_xxx33483';
             const result = database.markTaskComplete('nonexistent', 'user', documentName);
             expect(result).toBe(false);
         });
 
-        it('should create a new document if one with the given name does not exist', () => {
-            const documentName = 'nonexistent';
-            const taskId = database.addTask('test task', documentName);
-            database.addUserToTask(taskId, 'user', documentName);
-            const result = database.markTaskComplete(taskId, 'user', documentName);
-            expect(database.getDocument(documentName)).toBeDefined();
-            expect(result).toBe(true);
-        });
+
     });
 });
